Add sql statements for loading quizzes

+++ QuizWebsite/QuizWebsite/Pages/Quiz.cs
@@ -3,7 +3,7 @@
     public class Quiz
     {
 
-        public int QuizId { get; set; }
+        public long QuizId { get; set; }
 
         public string Title { get; set; }
